EfficientNet_inference/requirements.txt
tensorflow
numpy
pillow
uvicorn
fastapi
python-multipart
pydantic
Pillow
keras==3.10.0
tf2onnx
onnx
onnxruntime